Crate heapless_bytes

heapless-bytes

Newtype around heapless byte Vec with efficient serde.

Modules

consts
Type aliases for many constants.

Structs

Bytes

Traits

ArrayLength
Trait making GenericArray work, marking types to be used as length of an array

IsGreaterOrEqual
A type operator that returns True if Self >= Rhs, otherwise returns False.

Unsigned
The marker trait for compile time unsigned integers.

Type Definitions

Bytes8

Bytes16

Bytes32

Bytes64

True