Struct heapless_bytes::Bytes

pub struct Bytes<const N: usize> { /* fields omitted */ }

Implementations

impl<const N: usize> Bytes<N>

pub fn new() -> Self

Construct a new, empty Bytes<N>.

pub fn from<T: Into<Vec<u8, N>>>(bytes: T) -> Self

Wrap existing bytes in a Bytes<N>.

pub fn into_inner(self) -> Vec<u8, N>

Unwraps the Vec<u8, N>, same as into_vec.

pub fn into_vec(self) -> Vec<u8, N>

Unwraps the Vec<u8, N>, same as into_inner.

pub fn as_slice(&self) -> &[u8]

Returns an immutable slice view.

pub fn as_mut_slice(&mut self) -> &mut [u8]

Returns a mutable slice view.

pub fn try_convert_into<const M: usize>(&self) -> Result<Bytes<M>, ()>

Low-noise conversion between lengths.

We can’t implement TryInto since it would clash with blanket implementations.

pub fn from_slice(slice: &[u8]) -> Result<Self, ()>

pub fn try_from<E>(
f: impl FnOnce(&mut [u8]) -> Result<usize, E>
) -> Result<Self, E>

Some APIs offer an interface of the form f(&mut [u8]) -> Result<usize, E>, with the contract that the Ok-value signals how many bytes were written.

This constructor allows wrapping such interfaces in a more ergonomic way, returning a Bytes willed using f.

It seems it’s not possible to do this as an actual TryFrom implementation.

pub fn insert_slice_at(&mut self, slice: &[u8], at: usize) -> Result<(), ()>

pub fn insert(&mut self, index: usize, item: u8) -> Result<(), u8>

pub fn remove(&mut self, index: usize) -> Result<u8, ()>

pub fn resize_default(&mut self, new_len: usize) -> Result<(), ()>

pub fn resize_to_capacity(&mut self)

pub fn to_bytes<const M: usize>(&self) -> Result<Bytes<M>, ()>

Fallible conversion into differently sized byte buffer.

pub fn from_serialized<T>(t: &T) -> Self where
T: Serialize,